+++ data/nvim/lua/plugin/lsp_saga_c.lua
@@ -1,0 +1,4 @@
+local config = require('lspsaga').config
+config.lightbulb.enable = false;
+config.lightbulb.sign = false;
+config.lightbulb.virtual_text = false;